Revert "mac80211: QoS related cleanups"
This reverts commit 250bf595b81eac1ce99fe498e5bc6e93e41546aa.

+++ drivers/net/wireless/rt2x00/rt2x00.h
@@ -1002,7 +1002,7 @@ void rt2x00mac_bss_info_changed(struct ieee80211_hw *hw,
 				struct ieee80211_vif *vif,
 				struct ieee80211_bss_conf *bss_conf,
 				u32 changes);
-int rt2x00mac_conf_tx(struct ieee80211_hw *hw, u16 queue,
+int rt2x00mac_conf_tx(struct ieee80211_hw *hw, int queue,
 		      const struct ieee80211_tx_queue_params *params);
 
 /*
